removed duplicate or unsused functions
