+++ internal/references.go
@@ -33,11 +33,6 @@ func CreateETag(name string, size, timestamp int64) string {
 	return id.Fingerprint(fmt.Sprintf("%s%d%d", name, size, timestamp))
 }
 
-func CreateSimpleID() string {
-	id, _ := id.ShortUUID()
-	return id
-}
-
 func CreateSimpleToken() string {
 	token, _ := id.UUID()
 	return token
